+++ pipettes/simulator/CMakeLists.txt
@@ -28,20 +28,11 @@ add_library(freertos-pipettes STATIC ${FREERTOS_SOURCES})
 target_include_directories(freertos-pipettes PUBLIC ${PIPETTE_SIM_INCLUDE})
 
 function(add_simulator_executable TARGET)
-    if(${TARGET} STREQUAL "pipettes-single-simulator")
-        set(PIPETTES_SIMULATOR_SRC
-                ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/freertos_idle_timer_task.cpp
-                )
-    else()
-        set(PIPETTES_SIMULATOR_SRC
-                ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/freertos_idle_timer_task.cpp
-                )
-    endif()
     add_executable(
             ${TARGET}
-            ${PIPETTES_SIMULATOR_SRC}
+            ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/freertos_idle_timer_task.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/interfaces.cpp
             )
     target_include_directories(${TARGET} PUBLIC ${PIPETTE_SIM_INCLUDE})
     target_i2c_simlib(${TARGET})
